Moved function registration and other function-related methods to be accessible from DistributedTaskManager. Added FunctionType to DistributedFunction to enable function config when creating a distributed function.

+++ examples/simulator/distributionComplex/configProtocol.hpp
@@ -74,7 +74,7 @@ public:
                 return false;
             }
 
-            auto packet = PBuf::allocate( _state.currentModuleState().size() + sizeof( int ) );
+            auto packet = PBuf::allocate( static_cast< int >( _state.currentModuleState().size() + sizeof( int ) ) );
             auto payload = packet.payload();
             as< int >( payload ) = getConnectorId( interface.name() );
             auto moduleStateBuffer = payload + sizeof( int );
